requirements.txt
# Core dependencies
python-dateutil>=2.8.2
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyYAML>=6.0

# HTTP & Web
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# RSS
feedparser>=6.0.10

# LLM
openai>=1.12.0

# Email
Jinja2>=3.1.0

# Scheduling
APScheduler>=3.10.0
